+++ pkg/provider/nacos/discovery.go
@@ -1,46 +1,26 @@
 package nacos
 
 import (
+	"context"
 	"fmt"
 	"net/url"
 	"strings"
-	"time"
 
 	"github.com/nacos-group/nacos-sdk-go/v2/clients"
-	"github.com/nacos-group/nacos-sdk-go/v2/clients/config_client"
+	"github.com/nacos-group/nacos-sdk-go/v2/clients/naming_client"
 	"github.com/nacos-group/nacos-sdk-go/v2/common/constant"
+	"github.com/nacos-group/nacos-sdk-go/v2/model"
 	"github.com/nacos-group/nacos-sdk-go/v2/vo"
 	"github.com/nite-coder/bifrost/pkg/provider"
 	"github.com/nite-coder/blackbear/pkg/cast"
 )
 
-type Options struct {
-	Username    string
-	Password    string
-	NamespaceID string
-	Prefix      string
-	LogDir      string
-	LogLevel    string // debug, info, warn, error
-	CacheDir    string
-	Timeout     time.Duration
-	Watch       bool
-	Endpoints   []string
-	Files       []*File
+type NacosServiceDiscovery struct {
+	client  naming_client.INamingClient
+	options *Options
 }
 
-type File struct {
-	DataID  string
-	Group   string
-	Content string
-}
-
-type NacosProvider struct {
-	client    config_client.IConfigClient
-	options   Options
-	OnChanged provider.ChangeFunc
-}
-
-func NewProvider(options Options) (*NacosProvider, error) {
+func NewNacosServiceDiscovery(options Options) (*NacosServiceDiscovery, error) {
 	serverConfigs := []constant.ServerConfig{}
 
 	contextPath := "/nacos"
@@ -101,7 +81,7 @@ func NewProvider(options Options) (*NacosProvider, error) {
 		constant.WithPassword(options.Password),
 	)
 
-	client, err := clients.NewConfigClient(
+	client, err := clients.NewNamingClient(
 		vo.NacosClientParam{
 			ClientConfig:  &configConfig,
 			ServerConfigs: serverConfigs,
@@ -111,63 +91,50 @@ func NewProvider(options Options) (*NacosProvider, error) {
 		return nil, err
 	}
 
-	return &NacosProvider{
-		options: options,
+	return &NacosServiceDiscovery{
 		client:  client,
+		options: &options,
 	}, nil
+
 }
 
-func (p *NacosProvider) SetOnChanged(changeFunc provider.ChangeFunc) {
-	p.OnChanged = changeFunc
+func (d *NacosServiceDiscovery) GetInstances(ctx context.Context, options provider.GetInstanceOptions) ([]provider.Instancer, error) {
+	nacosInstances, err := d.client.SelectInstances(vo.SelectInstancesParam{
+		ServiceName: options.ID,
+		GroupName:   options.Group,
+		HealthyOnly: true,
+	})
+
+	if err != nil {
+		if err.Error() == "instance list is empty!" {
+			return nil, nil
+		}
+		return nil, fmt.Errorf("fail to select instances from nacos, error: %w, discovery id: %s, group: %s", err, options.ID, options.Group)
+	}
+
+	instances := ToProviderInstance(nacosInstances)
+	return instances, nil
 }
 
-func (p *NacosProvider) ConfigOpen() ([]*File, error) {
-	result := make([]*File, 0, len(p.options.Files))
+func (d *NacosServiceDiscovery) Watch(ctx context.Context, options provider.GetInstanceOptions) (<-chan []provider.Instancer, error) {
+	ch := make(chan []provider.Instancer, 1)
 
-	for _, file := range p.options.Files {
+	err := d.client.Subscribe(&vo.SubscribeParam{
+		ServiceName: options.ID,
+		GroupName:   options.Group,
+		SubscribeCallback: func(nacosInstances []model.Instance, err error) {
+			if err != nil {
+				return
+			}
 
-		group := "DEFAULT_GROUP"
-		if file.Group != "" {
-			group = file.Group
-		}
+			instances := ToProviderInstance(nacosInstances)
+			ch <- instances
+		},
+	})
 
-		content, err := p.client.GetConfig(vo.ConfigParam{
-			DataId: file.DataID,
-			Group:  file.Group,
-		})
-		if err != nil {
-			return nil, fmt.Errorf("nacos: fail to get '%s' file from '%s' group in '%s' namespace_id, error: %w", file.DataID, file.Group, p.options.NamespaceID, err)
-		}
-
-		newFile := &File{
-			DataID:  file.DataID,
-			Group:   group,
-			Content: content,
-		}
-
-		result = append(result, newFile)
-	}
-	return result, nil
-}
-
-func (p *NacosProvider) Watch() error {
-	if !p.options.Watch {
-		return nil
+	if err != nil {
+		return nil, err
 	}
 
-	for _, file := range p.options.Files {
-		err := p.client.ListenConfig(vo.ConfigParam{
-			DataId: file.DataID,
-			Group:  file.Group,
-			OnChange: func(namespace, group, dataId, data string) {
-				_ = p.OnChanged()
-			},
-		})
-
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
+	return ch, nil
 }